requirements.txt
earthengine-api>=0.1.332
geemap>=0.20.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
rasterio>=1.2.0
requests>=2.26.0
pyyaml>=6.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
geopandas>=0.10.0
xarray>=0.20.0
netCDF4>=1.5.0
plotly>=5.5.0
jupyterlab>=3.2.0
pyproj>=3.3.0
tqdm>=4.62.0